# contracts/iot-ingestion.openapi.yaml
openapi: 3.1.0
info:
  title: Smart Campus IoT Ingestion API
  version: 1.0.0
  description: |
    API và event contract cho IoT Ingestion service.
    REST paths dùng để ghi nhận, truy vấn và backfill dữ liệu IoT; hợp đồng chính của Pair-05 và Pair-06 là Queue async event.
    Pair-05: IoT → Core Business (sensor events)
    Pair-06: IoT → Analytics (telemetry events)
  contact:
    name: IoT Team
    email: iot@campus.local
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://iot.campus.local
    description: IoT Ingestion service
...
security:
  - bearerAuth: []
paths:
...
  /readings:
    post:
      operationId: createLabReading
      summary: Gửi một sensor reading vào stack Lab 05
      tags: [lab-readings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabSensorReadingCreate'
      responses:
        '201':
          description: Reading đã được AI service xử lý và lưu vào PostgreSQL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabSensorReadingCreated'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          description: Database hoặc AI service chưa sẵn sàng
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /readings/latest:
    get:
      operationId: getLatestLabReadings
      summary: Lấy các reading mới nhất đã lưu trong PostgreSQL
      tags: [lab-readings]
      parameters:
        - name: device_id
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Danh sách reading mới nhất
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/LabSensorReading'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /readings/{reading_id}:
    get:
      operationId: getLabReadingById
      summary: Lấy một reading theo ID
      tags: [lab-readings]
      parameters:
        - name: reading_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^R-[0-9]{8}-[0-9]{4}$'
      responses:
        '200':
          description: Chi tiết reading
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabSensorReading'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    LabSensorReadingCreate:
      type: object
      additionalProperties: false
      required: [device_id, metric, value, timestamp]
      properties:
        device_id:
          type: string
          minLength: 3
          example: ESP32-LAB-A01
        metric:
          type: string
          enum: [temperature, humidity, motion, smoke]
        value:
          type: number
          minimum: -40
          maximum: 80
          example: 31.5
        unit:
          type: string
          enum: [celsius, percent, boolean, ppm]
        timestamp:
          type: string
          format: date-time
    LabSensorReading:
      allOf:
        - $ref: '#/components/schemas/LabSensorReadingCreate'
        - type: object
          required: [reading_id, created_at]
          properties:
            reading_id:
              type: string
              pattern: '^R-[0-9]{8}-[0-9]{4}$'
            created_at:
              type: string
              format: date-time
    LabSensorReadingCreated:
      type: object
      additionalProperties: false
      required: [reading_id, device_id, metric, accepted, created_at]
      properties:
        reading_id:
          type: string
          pattern: '^R-[0-9]{8}-[0-9]{4}$'
        device_id:
          type: string
        metric:
          type: string
          enum: [temperature, humidity, motion, smoke]
        accepted:
          type: boolean
          const: true
        created_at:
          type: string
          format: date-time
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: URI định danh loại lỗi
          example: https://iot.campus.local/errors/validation
        title:
          type: string
          minLength: 3
          maxLength: 120
          description: Tiêu đề lỗi ngắn gọn
          example: Dữ liệu không hợp lệ
        status:
          type: integer
          minimum: 400
          maximum: 599
          description: HTTP status code
          example: 400
        detail:
          anyOf:
            - type: string
              maxLength: 500
            - type: 'null'
          description: Mô tả chi tiết lỗi
          example: Field deviceId không đúng định dạng
        instance:
          anyOf:
            - type: string
              format: uri
            - type: 'null'
          description: URI của request gây lỗi
          example: https://iot.campus.local/sensors/readings
        errors:
          type: array
          minItems: 0
          maxItems: 50
          items:
            type: object
            additionalProperties: false
            required: [field, code, message]
            properties:
              field:
                type: string
                minLength: 1
                maxLength: 80
                example: deviceId
              code:
                type: string
                minLength: 2
                maxLength: 80
                example: PATTERN_MISMATCH
              message:
                type: string
                minLength: 3
                maxLength: 200
                example: deviceId phải có dạng [A-Z0-9-]+
  responses:
...
    Unauthorized:
      description: Thiếu hoặc sai thông tin xác thực
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://iot.campus.local/errors/unauthorized
            title: Chưa xác thực
            status: 401
            detail: Thiếu Bearer token
            instance: https://iot.campus.local/sensors/readings
            errors: []
    NotFound:
      description: Resource không tồn tại
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://iot.campus.local/errors/not-found
            title: Không tìm thấy
            status: 404
            detail: Device không tồn tại
            instance: https://iot.campus.local/devices/SENSOR-999/status
            errors: []
    UnprocessableEntity:
      description: Payload đúng định dạng nhưng không hợp lệ về nghiệp vụ
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://iot.campus.local/errors/business-rule
            title: Dữ liệu không hợp lệ về nghiệp vụ
            status: 422
            detail: timestamp của sensor reading nằm ngoài khoảng thời gian cho phép
            instance: https://iot.campus.local/sensors/readings
            errors:
              - field: timestamp
                code: OUT_OF_RANGE
                message: timestamp không được lớn hơn thời điểm hiện tại quá 1 phút
...